requirements.txt
wyoming>=1.5.0
aiohttp>=3.9.0
pyaudio>=0.2.14
onnxruntime>=1.16.0
numpy>=1.24.0
tqdm>=4.0.0
requests>=2.0.0
scipy>=1.3.0
scikit-learn>=1.0.0
python-dotenv>=1.0.0
